readme
# Eletronic_voting-E-lectoral_System
This is a simple electronic voting system which create a platform for registration and casting of vote. 
to run this application on your local system, download xamp or wamp paste the the project profile on htdocs in your xampp file. upload the database phonebooks.sql on any sql database system. The Username and password of the registration portal is "admin"